requirements.txt
numpy>=1.21.0
pandas>=1.3.0
librosa>=0.9.0
soundfile>=0.11.0
scipy>=1.7.0
scikit-learn>=1.0.0
nltk>=3.6.0
spacy>=3.0.0
openai-whisper>=20230314
SpeechRecognition>=3.10.0
matplotlib>=3.4.0
seaborn>=0.11.0
pydub>=0.25.1
